5 - [Tech] Modules Management - remove previously added modules and move its contents to directories they're belongs to
