+++ app/src/main/java/io/minoro75/demoweatherapp/data/remote/RemoteDataSourceImpl.kt
@@ -1,6 +1,6 @@
-package io.minoro75.demoweatherapp.remote
+package io.minoro75.demoweatherapp.data.remote
 
-import io.minoro75.demoweatherapp.forecast_entity.ForecastLocation
+import io.minoro75.demoweatherapp.domain.forecasts.model.ForecastLocation
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
@@ -10,7 +10,7 @@ class RemoteDataSourceImpl @Inject constructor(
 ) : RemoteDataSource {
     override suspend fun getWeatherInCity(city: String): Flow<ForecastLocation> {
         return flow {
-            emit(remoteService.getWeatherInCity(city).forecasts?.forecastLocation!!)
+            emit(remoteService.getWeatherInCity(city).forecasts.forecastLocation)
         }
     }
