+++ app/src/main/java/io/minoro75/demoweatherapp/ui/city_selection/CitySelectionViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.minoro75.demoweatherapp.repository.WeatherRepository
+import io.minoro75.demoweatherapp.data.repository.WeatherRepository
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import javax.inject.Inject
