--- domain/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'kotlin'
-    id 'kotlin-kapt'
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_7
-    targetCompatibility = JavaVersion.VERSION_1_7
-}
-dependencies {
-
-    // coroutines
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
-    // moshi
-    implementation "com.squareup.moshi:moshi:1.12.0"
-    kapt "com.squareup.moshi:moshi-kotlin-codegen:1.12.0"
-}
